Move: entities, repositories 폴더로 entity, repository 이동
